CI(test): add custom proxy to playwright-actions
This changes the playwright action to use a new custom testing proxy,
which should make tests faster and more stable.
For the custom proxy a new config file with routes was added.
Also resets the sharding of PW tests as in our case one larger worker
seems to be a better choice.