Test: migrate integration test from a separate repo
This migrates the stuff from our playwright/integration test repo into
this one so it's easier to use and less repos to worry about.
Slightly tweaks some stuff from that repo so it works easier together
with the stuff in this one.
Also removes currently unused scripts and changes what tests are going
to be run as nightly/integration, simplifies this setup for now, we can
add stuff back later if desired.