feat(ci): add malicious package check
This adds the "dependency-review-action" to our PR CI workflow, which
will check our dependencies for any malicious packages.
Also adds a script that fetches the open-source "Aikido" malware
list and adds those to the review action as packages which should be
denied.
It would be good to have something like this with all the recent NPM
supply-chain attacks.