test: Check number of erratas not just the alphabetically-last
`dnf updateinfo --list --all` shows all advisories, even on installed
and older versions of packages. That means, if there is older errata and
there is no update that would alphabetically be "newer" then it fails.
Using count will compare how many advisories existed on the old
snapshot and how many exist on the new one. (OOI, on the current RHEL-9
old snapshots there are ~700 and on "latest" there are ~1200)

Also removed unnecessary subman refresh. We should not need it.